pyproject.toml
[project]
name = "technical-test-ahtglobal"
version = "0.1.0"
description = ""
authors = [
    {name = "JeanPaulSB",email = "jeanpaulsierraboom@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "mysql-connector-python (>=9.3.0,<10.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "cryptography (>=45.0.3,<46.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
